requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cloud-sql-python-connector[pymysql]
python-dotenv